backend for react project

+++ app/Http/Controllers/Admin/CarController.php
@@ -13,19 +13,19 @@ class CarController extends Controller
 {
     public function add_car(Request $request)
     {
-        $request->validate([
-            'car_name'=>'required',
-            'car_model'=>'required',
-            'car_type'=>'required',
-            'car_number'=>'required',
-            'rent_price'=>'required',
-            'car_buy_date'=>'required',
-            'car_color'=>'required',
-            'sit_number'=>'required',
-            'car_details'=>'required',
-            'car_owner_username'=>'required',
-            'car_owner_id'=>'required',
-          ]);
+        // $request->validate([
+        //     'car_name'=>'required',
+        //     'car_model'=>'required',
+        //     'car_type'=>'required',
+        //     'car_number'=>'required',
+        //     'rent_price'=>'required',
+        //     'car_buy_date'=>'required',
+        //     'car_color'=>'required',
+        //     'sit_number'=>'required',
+        //     'car_details'=>'required',
+        //     'car_owner_username'=>'required',
+        //     'car_owner_id'=>'required',
+        //   ]);
           $car= new CarService();
           $car->car_name=$request->car_name;
           $car->car_model=$request->car_model;
@@ -38,17 +38,18 @@ class CarController extends Controller
           $car->rent_price=$request->rent_price;
           $car->car_owner_name=$request->car_owner_username;
           $car->owner_id=$request->car_owner_id;
-          if($request->car_pic==null)
-          {
-            $car->car_pic= "carrental.png";
+          $car->rent_status=0;
+        //   if($request->car_pic==null)
+        //   {
+        //     $car->car_pic= "carrental.png";
 
-          }
-          else
-          {
-            $file_name = time().".".$request->file('car_pic')->getClientOriginalExtension();
-            $request->file('car_pic')->move(public_path('car_pic'),$file_name);
-            $car->car_pic = $file_name;
-          }
+        //   }
+        //   else
+        //   {
+        //     $file_name = time().".".$request->file('car_pic')->getClientOriginalExtension();
+        //     $request->file('car_pic')->move(public_path('car_pic'),$file_name);
+        //     $car->car_pic = $file_name;
+        //   }
           $res=$car->save();
           if($res){
             return 'You Add Car successfully';
@@ -78,17 +79,17 @@ class CarController extends Controller
         $car->car_owner_name=$request->car_owner_username;
         $car->owner_id=$request->car_owner_id;
 
-        if($request->car_pic==null)
-          {
-            $car->car_pic= $car->car_pic;
+        // if($request->car_pic==null)
+        //   {
+        //     $car->car_pic= $car->car_pic;
 
-          }
-          else
-          {
-            $file_name = time().".".$request->file('car_pic')->getClientOriginalExtension();
-            $request->file('car_pic')->move(public_path('car_pic'),$file_name);
-            $car->car_pic = $file_name;
-          }
+        //   }
+        //   else
+        //   {
+        //     $file_name = time().".".$request->file('car_pic')->getClientOriginalExtension();
+        //     $request->file('car_pic')->move(public_path('car_pic'),$file_name);
+        //     $car->car_pic = $file_name;
+        //   }
         $res=$car->save();
           if($res){
             return 'You Car Update successfully';
@@ -100,12 +101,17 @@ class CarController extends Controller
     public function delete_car(Request $req)
     {
         $car_edit=CarService::where('id','=',($req->id))->first();
-        $car_edit->delete();
-        $res=$car_edit;
+        if($car_edit){
+          $car_edit->delete();
+          $res=$car_edit;
           if($res){
             return 'You Car Delete successfully';
-        }else{
-            return 'Something Went Wrong';
+          }else{
+              return 'Something Went Wrong';
+          }
+        }
+        else{
+          return('Not found');
         }
     }
     public function search_car(Request $req)
